ZPI-40-manager-can-configure-redirect-uri
 - add contract
 - move api to interfaces
