+++ src/main/java/com/zpi/interfaces/rest/common/utils/HashGenerator.java
@@ -1,4 +1,4 @@
-package com.zpi.api.common.utils;
+package com.zpi.interfaces.rest.common.utils;
 
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
