+++ src/main/java/com/zpi/interfaces/rest/manager/ManagerController.java
@@ -1,8 +1,7 @@
-package com.zpi.api.manager;
+package com.zpi.interfaces.rest.manager;
 
-import com.zpi.api.manager.dto.ManagerDTO;
-import com.zpi.api.manager.dto.SignInDTO;
-import com.zpi.api.manager.dto.SignUpDTO;
+import com.zpi.interfaces.rest.manager.dto.SignInDTO;
+import com.zpi.interfaces.rest.manager.dto.SignUpDTO;
 import com.zpi.domain.manager.ManagerAccountManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
