+++ src/main/java/com/zpi/interfaces/rest/manager/dto/SignInDTO.java
@@ -1,4 +1,4 @@
-package com.zpi.api.manager.dto;
+package com.zpi.interfaces.rest.manager.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ManagerDTO {
+public class SignInDTO {
     private String username;
     private String password;
 }
