+++ src/main/java/com/zpi/interfaces/rest/manager/dto/SignUpDTO.java
@@ -1,0 +1,14 @@
+package com.zpi.interfaces.rest.manager.dto;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class SignUpDTO {
+    private String username;
+    private String password;
+}
