+++ src/main/resources/application.yaml
@@ -19,6 +19,7 @@ springdoc:
     path: /api/docs
 
 server:
+  port: 10000
   error:
     include-message: always
