+++ src/test/groovy/com/zpi/interfaces/rest/client/AddRedirectionURIFT.groovy
@@ -1,4 +1,4 @@
-package com.zpi.api.client
+package com.zpi.interfaces.rest.client
 
 import com.zpi.CommonFixtures
 import com.zpi.CommonHelpers
