+++ src/test/groovy/com/zpi/interfaces/rest/client/ClientRegistrationFT.groovy
@@ -1,4 +1,4 @@
-package com.zpi.api.client
+package com.zpi.interfaces.rest.client
 
 
 import com.zpi.CommonFixtures
